# openapi/assignment-api.yaml
openapi: 3.0.0
info:
  title: NS Assignement API - By Arjen Klaverstijn
  version: 1.0.0
paths:
  /api/v3/comfort:
    get:
      summary: Get the trips sorted by most comfortable!
      operationId: getMostComfortable
      parameters:
        - name: fromStation
          in: query
          description: NS station code of the origin station
          required: true  
          schema:
            type: string
        - name: toStation
          in: query
          description: NS station code of the destination station
          required: true
          schema:
            type: string
        - name: dateTime
          in: query
          description: Format - date-time (as date-time in RFC3339). Datetime that the
            user want to depart from his origin or or arrive at his destination default is server time
          schema:
            type: string
            format: date-time

      responses:
        '200':
          description: A list of trips sorted by comfort
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Trip'
        '404':
          description: Could not find trips
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '500':  
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '429':  
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
  /api/v3/optimal:
    get:
      summary: Get the the most optimal trip!
      operationId: getMostOptimal
      parameters:
        - name: fromStation
          in: query
          description: NS station code of the origin station
          required: true  
          schema:
            type: string
        - name: toStation
          in: query
          description: NS station code of the destination station
          required: true  
          schema:
            type: string
        - name: dateTime
          in: query
          description: Format - date-time (as date-time in RFC3339). Datetime that the
            user want to depart from his origin or or arrive at his destination default is server time
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: return the trip that the ns finds most optimal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trip'
        '404':
          description: Could not find optimal trip
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '500':  
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '429':  
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
components:
  schemas:
    Coordinate:
      required:
        - lat
        - lng
      type: object
      properties:
        lat:
          type: number
        lng:
          type: number
    Duration:
      required:
        - accessibilityValue
        - nesProperties
        - value
      type: object
      properties:
        value:
          type: string
        accessibilityValue:
          type: string
        nesProperties:
          $ref: '#/components/schemas/NesProperties'
    Eco:
      required:
        - co2kg
      type: object
      properties:
        co2kg:
          type: number
          format: double
    EticketNotBuyableReason:
      required:
        - reason
      type: object
      properties:
        reason:
          enum:
            - UNKNOWN_PRICE
            - TOO_MANY_SEPARATE_PARTS
            - TOO_FAR_IN_PAST
            - TOO_FAR_IN_FUTURE
            - STATION_NOT_OPEN_YET
            - TRIP_IS_NOT_DOMESTIC
            - VIA_STATION_REQUESTED
            - NO_TRAIN_LEGS_IN_TRIP
          type: string
        description:
          type: string
    FareLeg:
      required:
        - destination
        - fares
        - origin
        - productTypes
      type: object
      properties:
        origin:
          $ref: '#/components/schemas/TripOriginDestination'
        destination:
          $ref: '#/components/schemas/TripOriginDestination'
        operator:
          type: string
        productTypes:
          type: array
          items:
            enum:
              - TRAIN
              - BUS
              - TRAM
              - METRO
              - FERRY
              - WALK
              - BIKE
              - CAR
              - TAXI
              - SHARED_MODALITY
              - UNKNOWN
            type: string
        fares:
          type: array
          items:
            $ref: '#/components/schemas/TripTravelFare'
    FareLegStop:
      required:
        - varCode
      type: object
      properties:
        varCode:
          type: integer
          format: int32
        name:
          type: string
    FareRoute:
      required:
        - destination
        - origin
      type: object
      properties:
        routeId:
          type: string
        origin:
          $ref: '#/components/schemas/FareLegStop'
        destination:
          $ref: '#/components/schemas/FareLegStop'
    JourneyDetailLink:
      required:
        - link
        - type
      type: object
      properties:
        type:
          enum:
            - BTM
            - TRAIN_XML
          type: string
        link:
          $ref: '#/components/schemas/Link'
    JourneyRegistrationParameters:
      required:
        - bicycleReservationRequired
        - searchUrl
        - status
      type: object
      properties:
        url:
          type: string
          format: uri
        searchUrl:
          type: string
          format: uri
        status:
          enum:
            - REGISTRATION_POSSIBLE
            - NOT_AVAILABLE
            - DATE_IN_PAST
            - DATE_TOO_FAR_FUTURE
            - NOT_NECESSARY_OTHER_OPERATOR
            - UNKNOWN
          type: string
        bicycleReservationRequired:
          type: boolean
        availability:
          $ref: '#/components/schemas/RegistrationAvailability'
    LabelListItem:
      required:
        - label
        - stickerType
      type: object
      properties:
        label:
          type: string
        stickerType:
          type: string
        salesOptionType:
          enum:
            - NS_DEAL_DISCOUNT
          type: string
    Leg:
      required:
        - alternativeTransport
        - cancelled
        - changePossible
        - destination
        - isOnOrAfterCancelledLeg
        - origin
        - partCancelled
        - postSteps
        - preSteps
        - reachable
        - stops
      type: object
      properties:
        idx:
          type: string
        name:
          type: string
        travelType:
          enum:
            - PUBLIC_TRANSIT
            - WALK
            - TRANSFER
            - BIKE
            - CAR
            - KISS
            - TAXI
            - UNKNOWN
          type: string
        direction:
          type: string
        partCancelled:
          type: boolean
        cancelled:
          type: boolean
        isOnOrAfterCancelledLeg:
          type: boolean
        changePossible:
          type: boolean
        alternativeTransport:
          type: boolean
        journeyDetailRef:
          type: string
        origin:
          $ref: '#/components/schemas/TripOriginDestination'
        destination:
          $ref: '#/components/schemas/TripOriginDestination'
        product:
          $ref: '#/components/schemas/ProductInterface'
        sharedModality:
          $ref: '#/components/schemas/SharedModality'
        notes:
          type: array
          items:
            $ref: '#/components/schemas/Note'
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        transferMessages:
          type: array
          items:
            $ref: '#/components/schemas/TransferMessage'
        stops:
          type: array
          items:
            $ref: '#/components/schemas/Stop'
        steps:
          type: array
          items:
            $ref: '#/components/schemas/Step'
        coordinates:
          type: array
          items:
            type: array
            items:
              type: number
              format: double
        crowdForecast:
          enum:
            - UNKNOWN
            - LOW
            - MEDIUM
            - HIGH
          type: string
        bicycleSpotCount:
          type: integer
          format: int32
        punctuality:
          type: number
          format: double
        crossPlatformTransfer:
          type: boolean
        shorterStock:
          type: boolean
        changeCouldBePossible:
          type: boolean
        shorterStockWarning:
          type: string
        shorterStockClassification:
          enum:
            - BUSY
            - EXTRA_BUSY
          type: string
        journeyDetail:
          type: array
          items:
            $ref: '#/components/schemas/JourneyDetailLink'
        reachable:
          type: boolean
        plannedDurationInMinutes:
          type: integer
          format: int64
        travelAssistanceDeparture:
          $ref: '#/components/schemas/ServiceBookingInfo'
        travelAssistanceArrival:
          $ref: '#/components/schemas/ServiceBookingInfo'
        overviewPolyLine:
          type: array
          items:
            $ref: '#/components/schemas/Coordinate'
        nesProperties:
          $ref: '#/components/schemas/NesProperties'
        duration:
          $ref: '#/components/schemas/Duration'
        preSteps:
          type: array
          items:
            $ref: '#/components/schemas/Step'
        postSteps:
          type: array
          items:
            $ref: '#/components/schemas/Step'
        transferTimeToNextLeg:
          type: integer
          format: int64
    Link:
      required:
        - uri
      type: object
      properties:
        uri:
          type: string
    MeetingPointDetails:
      required:
        - minutesBefore
        - name
      type: object
      properties:
        name:
          type: string
        minutesBefore:
          type: integer
          format: int32
    Message:
      type: object
      properties:
        id:
          type: string
        externalId:
          type: string
        head:
          type: string
        text:
          type: string
        lead:
          type: string
        routeIdxFrom:
          type: integer
          format: int32
        routeIdxTo:
          type: integer
          format: int32
        type:
          enum:
            - MAINTENANCE
            - DISRUPTION
            - SHORTENED
          type: string
        nesProperties:
          $ref: '#/components/schemas/NesProperties'
        startDate:
          type: string
        endDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        phase:
          enum:
            - PHASE_1A
            - PHASE_1B
            - PHASE_2
            - PHASE_3
            - PHASE_4
            - PHASE_5
          type: string
    ModalityListItem:
      required:
        - accessibilityName
        - name
        - nameNesProperties
      type: object
      properties:
        name:
          type: string
        nameNesProperties:
          $ref: '#/components/schemas/NesProperties'
        iconNesProperties:
          $ref: '#/components/schemas/NesProperties'
        actualTrack:
          type: string
        accessibilityName:
          type: string
    NesIcon:
      required:
        - color
        - icon
      type: object
      properties:
        color:
          type: string
        icon:
          type: string
        backgroundColor:
          type: string
    NesLine:
      required:
        - color
      type: object
      properties:
        color:
          type: string
        styles:
          $ref: '#/components/schemas/NesLineStyles'
    NesLineStyles:
      required:
        - dashed
      type: object
      properties:
        dashed:
          type: boolean
    NesProperties:
      required:
        - color
      type: object
      properties:
        color:
          type: string
        type:
          type: string
        icon:
          type: string
        scope:
          enum:
            - LEG_LINE
            - ENTIRE_LEG
          type: string
        styles:
          $ref: '#/components/schemas/Styles'
    NesText:
      required:
        - accessibilityText
        - color
        - text
      type: object
      properties:
        color:
          type: string
        text:
          type: string
        accessibilityText:
          type: string
        backgroundColor:
          type: string
        styles:
          $ref: '#/components/schemas/NesTextStyles'
    NesTextStyles:
      required:
        - bold
        - strikethrough
      type: object
      properties:
        strikethrough:
          type: boolean
        bold:
          type: boolean
    Note:
      required:
        - isPresentationRequired
      type: object
      properties:
        value:
          type: string
        shortValue:
          type: string
        accessibilityValue:
          type: string
        key:
          type: string
        noteType:
          enum:
            - UNKNOWN
            - ATTRIBUTE
            - INFOTEXT
            - REALTIME
            - TICKET
            - HINT
          type: string
        priority:
          type: integer
          format: int32
        routeIdxFrom:
          type: integer
          format: int32
        routeIdxTo:
          type: integer
          format: int32
        link:
          $ref: '#/components/schemas/Link'
        isPresentationRequired:
          type: boolean
        category:
          enum:
            - PLATFORM_INFORMATION
            - OVERCHECK_INSTRUCTION
            - UNKNOWN
          type: string
        nesProperties:
          $ref: '#/components/schemas/NesProperties'
      description: notes is a List of Lists because this allows frontend clients to
        render the elements of every nested List on the same row
    NsiLink:
      required:
        - showInternationalBanner
        - url
      type: object
      properties:
        url:
          type: string
          description: NSI journeyplanner url with pre-filled locations and time
        showInternationalBanner:
          type: boolean
          description: Whether to display an NS International banner
    PrimaryMessage:
      required:
        - nesProperties
        - title
        - type
      type: object
      properties:
        title:
          type: string
        nesProperties:
          $ref: '#/components/schemas/NesProperties'
        message:
          $ref: '#/components/schemas/Message'
        type:
          enum:
            - TRIP_CANCELLED
            - LEG_CANCELLED
            - LEG_TRANSFER_IMPOSSIBLE
            - ALTERNATIVE_TRANSPORT
            - DISRUPTION
            - MAINTENANCE
            - REPLACEMENT
            - ADDITIONAL
            - SHORTENED_TRAIN
          type: string
      description: Most important message to display
...
    ProductInterface:
      required:
        - nameNesProperties
        - notes
        - productType
        - type
      type: object
      properties:
        number:
          type: string
        displayName:
          type: string
        operatorCode:
          type: string
        notes:
          type: array
          items:
            type: array
            items:
              $ref: '#/components/schemas/Note'
        operatorName:
          type: string
        nameNesProperties:
          $ref: '#/components/schemas/NesProperties'
        iconNesProperties:
          $ref: '#/components/schemas/NesProperties'
        categoryCode:
          type: string
        longCategoryName:
          type: string
        shortCategoryName:
          type: string
        operatorAdministrativeCode:
          type: integer
          format: int32
        type:
          enum:
            - TRAIN
            - BUS
            - TRAM
            - METRO
            - FERRY
            - WALK
            - BIKE
            - CAR
            - TAXI
            - SHARED_MODALITY
            - UNKNOWN
          type: string
        productType:
          type: string
      discriminator:
        propertyName: productType
    RegistrationAvailability:
      required:
        - bicycle
        - seats
      type: object
      properties:
        seats:
          type: boolean
        numberOfSeats:
          type: integer
          format: int32
        bicycle:
          type: boolean
        numberOfBicyclePlaces:
          type: integer
          format: int32
    SalesOption:
      required:
        - betterOption
        - type
      type: object
      properties:
        type:
          enum:
            - NS_DEAL_DISCOUNT
          type: string
        permilleFullTariff:
          type: integer
          format: int32
        priceInCents:
          type: integer
          format: int32
        originalPrice:
          type: integer
          format: int32
        betterOption:
          type: boolean
        recommendationText:
          type: string
        labelText:
          type: string
    ServiceBookingInfo:
      required:
        - canChangeAssistance
        - defaultAssistanceValue
        - name
        - serviceTypeIds
        - tripLegIndex
      type: object
      properties:
        name:
          type: string
        tripLegIndex:
          type: string
        stationUic:
          type: string
        serviceTypeIds:
          type: array
          items:
            type: string
        defaultAssistanceValue:
          type: boolean
        canChangeAssistance:
          type: boolean
        message:
          type: string
    SharedModality:
      required:
        - availability
        - nearByMeMapping
        - provider
      type: object
      properties:
        provider:
          type: string
        name:
          type: string
        availability:
          type: boolean
        nearByMeMapping:
          enum:
            - OV_FIETS
            - SHARED_ELECTRICAL_BIKE
            - SHARED_BIKE
            - SHARED_SCOOTER
            - SHARED_CAR
            - UNKNOWN
          type: string
        planIcon:
          type: string
    Step:
      required:
        - line
        - text
      type: object
      properties:
        text:
          $ref: '#/components/schemas/NesText'
        duration:
          $ref: '#/components/schemas/NesText'
        icon:
          $ref: '#/components/schemas/NesIcon'
        line:
          $ref: '#/components/schemas/NesLine'
    Stop:
      required:
        - borderStop
        - cancelled
        - notes
        - passing
      type: object
      properties:
        uicCode:
          type: string
        name:
          type: string
        lat:
          type: number
        lng:
          type: number
        countryCode:
          type: string
        notes:
          type: array
          items:
            $ref: '#/components/schemas/StopNote'
        routeIdx:
          type: integer
          format: int32
        departurePrognosisType:
          type: string
        plannedDepartureDateTime:
          type: string
          format: date-time
        plannedDepartureTimeZoneOffset:
          type: integer
          format: int32
        actualDepartureDateTime:
          type: string
          format: date-time
        actualDepartureTimeZoneOffset:
          type: integer
          format: int32
        plannedArrivalDateTime:
          type: string
          format: date-time
        plannedArrivalTimeZoneOffset:
          type: integer
          format: int32
        actualArrivalDateTime:
          type: string
          format: date-time
        actualArrivalTimeZoneOffset:
          type: integer
          format: int32
        plannedPassingDateTime:
          type: string
          format: date-time
        actualPassingDateTime:
          type: string
          format: date-time
        arrivalPrognosisType:
          type: string
        actualDepartureTrack:
          type: string
        plannedDepartureTrack:
          type: string
        plannedArrivalTrack:
          type: string
        actualArrivalTrack:
          type: string
        departureDelayInSeconds:
          type: integer
          format: int64
        arrivalDelayInSeconds:
          type: integer
          format: int64
        cancelled:
          type: boolean
        borderStop:
          type: boolean
        passing:
          type: boolean
        quayCode:
          type: string
    StopNote:
      required:
        - type
        - value
      type: object
      properties:
        value:
          type: string
        key:
          type: string
        type:
          enum:
            - U
            - A
            - I
            - R
            - H
          type: string
        priority:
          type: integer
          format: int32
    Styles:
      required:
        - type
      type: object
      properties:
        type:
          type: string
      discriminator:
        propertyName: type
    TransferMessage:
      required:
        - accessibilityMessage
        - message
        - messageNesProperties
        - type
      type: object
      properties:
        message:
          type: string
        accessibilityMessage:
          type: string
        type:
          enum:
            - IMPOSSIBLE_TRANSFER
            - SPECIAL
            - CHECKOVER_INSTRUCTION
            - CROSS_PLATFORM
            - TRANSFER_TIME
          type: string
        messageNesProperties:
          $ref: '#/components/schemas/NesProperties'
        iconNesProperties:
          $ref: '#/components/schemas/NesProperties'
...
    TravelAssistanceInfo:
      required:
        - isAssistanceRequired
        - tripRequestId
      type: object
      properties:
        termsAndConditionsLink:
          type: string
          format: url
        tripRequestId:
          type: integer
          format: int32
        isAssistanceRequired:
          type: boolean
    Trip:
      required:
        - ctxRecon
        - legs
        - optimal
        - realtime
        - status
        - transfers
        - type
        - uid
      type: object
      properties:
        uid:
          type: string
          description: Unique identifier for this trip
        ctxRecon:
          type: string
          description: Reconstruction context for this trip. Can be used to reconstruct
            this exact trip with the v3/trips/trip endpoint
        sourceCtxRecon:
          type: string
          description: Reconstruction context for this trip set by the travel advice
            source system. Can be used to interact follow-up requests with the
            source system, like booking a trip with travel assistance (NSRA).
            Can be null if the source doesn't provide a ctxRecon
        plannedDurationInMinutes:
          type: integer
          description: Planned duration of this trip in minutes
          format: int64
        actualDurationInMinutes:
          type: integer
          description: Actual duration of this trip in minutes, or the planned duration if
            no realtime information about this trip is available.
          format: int64
        transfers:
          type: integer
          description: Number of public transit transfers
          format: int32
        status:
          enum:
            - CANCELLED
            - CHANGE_NOT_POSSIBLE
            - ALTERNATIVE_TRANSPORT
            - DISRUPTION
            - MAINTENANCE
            - UNCERTAIN
            - REPLACEMENT
            - ADDITIONAL
            - SPECIAL
            - NORMAL
          type: string
          description: Status of this trip
        primaryMessage:
          $ref: '#/components/schemas/PrimaryMessage'
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
          description: List of messages regarding maintenance or disruption that
            influences this trip.
        legs:
          type: array
          items:
            $ref: '#/components/schemas/Leg'
        overviewPolyLine:
          type: array
          items:
            $ref: '#/components/schemas/Coordinate'
        crowdForecast:
          enum:
            - UNKNOWN
            - LOW
            - MEDIUM
            - HIGH
          type: string
        punctuality:
          type: number
          format: double
        optimal:
          type: boolean
          description: Whether or not this trip is regarded the best possible option of
            all returned trips
        fareRoute:
          $ref: '#/components/schemas/FareRoute'
        fares:
          type: array
          items:
            $ref: '#/components/schemas/TripSalesFare'
        fareLegs:
          type: array
          items:
            $ref: '#/components/schemas/FareLeg'
        productFare:
          $ref: '#/components/schemas/TripTravelFare'
        fareOptions:
          $ref: '#/components/schemas/TripFareOptions'
        bookingUrl:
          $ref: '#/components/schemas/Link'
        nsiLink:
          $ref: '#/components/schemas/NsiLink'
        type:
          enum:
            - NS
            - NS_ACCESSIBLE
            - NEGENTWEE
            - GOOGLE
            - PAS
          type: string
        shareUrl:
          $ref: '#/components/schemas/Link'
        realtime:
          type: boolean
        travelAssistanceInfo:
          $ref: '#/components/schemas/TravelAssistanceInfo'
        routeId:
          type: string
        registerJourney:
          $ref: '#/components/schemas/JourneyRegistrationParameters'
        eco:
          $ref: '#/components/schemas/Eco'
        modalityListItems:
          type: array
          items:
            $ref: '#/components/schemas/ModalityListItem'
        labelListItems:
          type: array
          items:
            $ref: '#/components/schemas/LabelListItem'
    TripFareOptions:
      required:
        - isEticketBuyable
        - isInternational
        - isInternationalBookable
        - isPossibleWithOvChipkaart
        - isTotalPriceUnknown
      type: object
      properties:
        isInternationalBookable:
          type: boolean
        isInternational:
          type: boolean
        isEticketBuyable:
          type: boolean
        isPossibleWithOvChipkaart:
          type: boolean
        isTotalPriceUnknown:
          type: boolean
        supplementsBasedOnSelectedFare:
          type: array
          items:
            $ref: '#/components/schemas/TripFareSupplement'
        reasonEticketNotBuyable:
          $ref: '#/components/schemas/EticketNotBuyableReason'
        salesOptions:
          type: array
          items:
            $ref: '#/components/schemas/SalesOption'
    TripFareSupplement:
      required:
        - supplementPriceInCents
      type: object
      properties:
        supplementPriceInCents:
          type: integer
          format: int32
        fromUICCode:
          type: string
        toUICCode:
          type: string
        link:
          $ref: '#/components/schemas/Link'
    TripOriginDestination:
      type: object
      properties:
        name:
          type: string
        lng:
          type: number
        lat:
          type: number
        city:
          type: string
        countryCode:
          type: string
        uicCode:
          type: string
        stationCode:
          type: string
        type:
          enum:
            - STATION
            - ADDRESS
            - POINT_OF_INTEREST
          type: string
        prognosisType:
          type: string
        plannedTimeZoneOffset:
          type: integer
          format: int32
        plannedDateTime:
          type: string
          format: date-time
        actualTimeZoneOffset:
          type: integer
          format: int32
        actualDateTime:
          type: string
          format: date-time
        plannedTrack:
          type: string
        actualTrack:
          type: string
        exitSide:
          enum:
            - LEFT
            - RIGHT
            - UNKNOWN
          type: string
        checkinStatus:
          enum:
            - CHECKIN
            - CHECKOUT
            - OVERCHECK
            - DETOUR
            - REQUIRED_CHECK_OUT_IN
            - NOTHING
          type: string
        travelAssistanceBookingInfo:
          $ref: '#/components/schemas/ServiceBookingInfo'
        travelAssistanceMeetingPoints:
          type: array
          items:
            type: string
        travelAssistanceMeetingPointDetails:
          type: array
          items:
            $ref: '#/components/schemas/MeetingPointDetails'
        notes:
          type: array
          items:
            $ref: '#/components/schemas/Note'
        quayCode:
          type: string
    TripSalesFare:
      type: object
      properties:
        priceInCents:
          type: integer
          format: int32
        product:
          enum:
            - OVCHIPKAART_ENKELE_REIS
            - OVCHIPKAART_RETOUR
            - TRAJECT_VRIJ_JAAR
            - BUSINESS_CARD_TRAJECT_VRIJ_JAAR
            - ETICKET_ENKELE_REIS
            - ETICKET_RETOUR
            - NS_DEAL_DISCOUNT_ENKELE_REIS
            - RAILRUNNER
            - ICE_SUPPLEMENT
            - ICD_SUPPLEMENT
            - NSI
          type: string
        travelClass:
          enum:
            - FIRST_CLASS
            - SECOND_CLASS
          type: string
        priceInCentsExcludingSupplement:
          type: integer
          format: int32
        discountType:
          enum:
            - NO_DISCOUNT
            - DISCOUNT_20_PERCENT
            - DISCOUNT_40_PERCENT
            - NO_CHARGE
            - OTHER
          type: string
        supplementInCents:
          type: integer
          format: int32
        link:
          type: string
    TripTravelFare:
      required:
        - discountType
      type: object
      properties:
        priceInCents:
          type: integer
          format: int32
        priceInCentsExcludingSupplement:
          type: integer
          format: int32
        supplementInCents:
          type: integer
          format: int32
        buyableTicketPriceInCents:
          type: integer
          format: int32
        buyableTicketPriceInCentsExcludingSupplement:
          type: integer
          format: int32
        buyableTicketSupplementPriceInCents:
          type: integer
          format: int32
        product:
          enum:
            - GEEN
            - OVCHIPKAART_ENKELE_REIS
            - OVCHIPKAART_RETOUR
            - DAL_VOORDEEL
            - ALTIJD_VOORDEEL
            - DAL_VRIJ
            - WEEKEND_VRIJ
            - ALTIJD_VRIJ
            - BUSINESSCARD
            - BUSINESSCARD_DAL
            - STUDENT_WEEK
            - STUDENT_WEEKEND
            - VDU
            - SAMENREISKORTING
            - TRAJECT_VRIJ
          type: string
        travelClass:
          enum:
            - FIRST_CLASS
            - SECOND_CLASS
          type: string
        discountType:
          enum:
            - NO_DISCOUNT
            - DISCOUNT_20_PERCENT
            - DISCOUNT_40_PERCENT
            - NO_CHARGE
            - OTHER
          type: string
        link:
          type: string
    APIError:
      required:
        - code
        - message
        - path
        - timestamp
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        code:
          type: integer
          format: int32
        message:
          type: string
        path:
          type: string
...
  securitySchemes:
    apiKeyHeader:
      type: apiKey
      name: Ocp-Apim-Subscription-Key
      in: header
    apiKeyQuery:
      type: apiKey
      name: subscription-key
      in: query